+++ R/get_join_keys.R
@@ -18,32 +18,32 @@ get_join_keys.TealData <- function(data) {
   data$get_join_keys()
 }
 
-#' @rdname get_join_keys
-#' @export
-get_join_keys.JoinKeys <- function(data) {
-  data
-}
-
-#' @rdname get_join_keys
-#' @export
-`get_join_keys<-` <- function(data, dataset_1, dataset_2 = NULL, value) {
-  UseMethod("get_join_keys<-", data)
-}
-
-#' @rdname get_join_keys
-#' @export
-`get_join_keys<-.JoinKeys` <- function(data, dataset_1, dataset_2 = NULL, value) {
-  data
-}
-
-#' @rdname get_join_keys
-#' @export
-`get_join_keys<-.teal_data` <- function(data, dataset_1, dataset_2 = NULL, value) {
-  data
-}
-
-#' @rdname get_join_keys
-#' @export
-`get_join_keys<-.TealData` <- function(data, dataset_1, dataset_2 = NULL, value) {
-  data
-}
+#' #' @rdname get_join_keys
+#' #' @export
+#' get_join_keys.JoinKeys <- function(data) {
+#'   data
+#' }
+#'
+#' #' @rdname get_join_keys
+#' #' @export
+#' `get_join_keys<-` <- function(data, dataset_1, dataset_2 = NULL, value) {
+#'   UseMethod("get_join_keys<-", data)
+#' }
+#'
+#' #' @rdname get_join_keys
+#' #' @export
+#' `get_join_keys<-.JoinKeys` <- function(data, dataset_1, dataset_2 = NULL, value) {
+#'   data
+#' }
+#'
+#' #' @rdname get_join_keys
+#' #' @export
+#' `get_join_keys<-.teal_data` <- function(data, dataset_1, dataset_2 = NULL, value) {
+#'   data
+#' }
+#'
+#' #' @rdname get_join_keys
+#' #' @export
+#' `get_join_keys<-.TealData` <- function(data, dataset_1, dataset_2 = NULL, value) {
+#'   data
+#' }